move persons modal logic out of trends
