+++ frontend/src/scenes/trends/viz/ActionsLineGraph.tsx
@@ -8,6 +8,7 @@ import { ACTIONS_BAR_CHART } from 'lib/constants'
 import { ChartParams } from '~/types'
 import { ViewType } from 'scenes/insights/insightLogic'
 import { router } from 'kea-router'
+import { personsModalLogic } from '../personsModalLogic'
 
 export function ActionsLineGraph({
     dashboardItemId,
@@ -24,7 +25,7 @@ export function ActionsLineGraph({
         cachedResults,
     })
     const { filters, indexedResults, resultsLoading, visibilityMap } = useValues(logic)
-    const { loadPeople } = useActions(logic)
+    const { loadPeople } = useActions(personsModalLogic)
     const [{ fromItem }] = useState(router.values.hashParams)
 
     return indexedResults && !resultsLoading ? (
